requirements.txt
ollama
openai
python-dotenv
colorama
